+++ packages/workflow/jest/file.js
@@ -4,6 +4,6 @@ const path = require('path');
 
 module.exports = {
   process(src, filename) {
-    return `module.exports = ${JSON.stringify(path.basename(filename))};`;
+    return { code: `module.exports = ${JSON.stringify(path.basename(filename))};` };
   }
 };