+++ packages/mock-server/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [7.2.4-alpha.0](https://github.com/Availity/availity-workflow/compare/@availity/mock-server@7.2.3...@availity/mock-server@7.2.4-alpha.0) (2023-07-28)
+
+
+### Build System
+
+* drop support for node 12 and node 14 ([6090a36](https://github.com/Availity/availity-workflow/commit/6090a36329132c09fadea261211395e2b87deb24))
+
+
+### BREAKING CHANGES
+
+* drop support for node 12 and node 14
+
+
+
 ## [7.2.3](https://github.com/Availity/availity-workflow/compare/@availity/mock-server@7.2.2...@availity/mock-server@7.2.3) (2023-03-17)